3.3.12

+++ src/main/java/me/shedaniel/rei/plugin/campfire/DefaultCampfireDisplay.java
@@ -45,7 +45,7 @@ public class DefaultCampfireDisplay implements RecipeDisplay {
         this.cookTime = cookTime;
     }
     
-    public int getCookTime() {
+    public double getCookTime() {
         return cookTime;
     }
